+++ classify/pre_data/pre_data.py
@@ -4,8 +4,13 @@ import torch.utils.data as Data
 import cv2
 import torch
 
+
 class GetData(Data.Dataset):
-    def __init__(self, path, ):
+
+    def __init__(
+        self,
+        path,
+    ):
         super(GetData, self).__init__()
         self.root = path
         self.all_files = os.listdir(self.root)
@@ -21,19 +26,3 @@ class GetData(Data.Dataset):
 
 
 
-class HandWrittenData(Data.Dataset):
-    def __init__(self, dir, transform = None):
-        super(HandWrittenData, self).__init__()
-        self.dir = dir
-        self.all_imgs = os.listdir(self.dir)
-        self.transform = transform
-    def __len__(self):
-        return len(self.all_imgs)
-
-    def __getitem__(self, item):
-        # img = cv2.imread(os.path.join(self.dir, self.all_imgs[item]), cv2.IMREAD_GRAYSCALE)
-        img = Image.open(os.path.join(self.dir, self.all_imgs[item])).convert("RGB")
-        if self.transform:
-            img = self.transform(img)
-        flag = int(self.all_imgs[item][4])
-        return img, torch.tensor(flag)
